CptnH00k now can accept hooks and continues them. The cool thing about this is that Cptnh00k makes webhooks transactional. It will keep retry- ing to deliver the hook until the application accepts it, or until the hooks TTL expires.
It also provides authentication for hooks, unlike with most webhook 
implementations, the target application can verify that the hook is 
being received from a valid source. H00k will also verify the source,
so there is a seamless authentication system going on.

To prevent too much stress on the source server, h00k also allows for 
Pre-Shared Keys to be exchanged. It will send one along with each 
request, to ensure that the receiving application can check the 
signature and therefore accept the request. It will also provide one to
every application sending requests.

This allows applications to verify future requests from and to h00k with
the PSK.